owl/pyproject.toml
[project]
name = "owl"
version = "0.1.0"
description = "owl by adam"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.3.0",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pyarrow>=21.0.0",
    "python-json-logger>=3.3.0",
    "rerun-sdk>=0.24.1",
    "typing-extensions>=4.15.0",
]
